+++ includes/frontend/class-menu.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  2.0.0
+ * @version  1.5.3
  *
  * Contents:
  *
@@ -438,7 +438,7 @@ class Reykjavik_Menu {
 			 * Social links supported icons.
 			 *
 			 * @since    1.0.0
-			 * @version  2.0.0
+			 * @version  1.5.3
 			 */
 			public static function social_links_icons() {