+++ templates/parts/admin/welcome-promo.php
@@ -26,9 +26,9 @@
 
 <div class="welcome-upgrade">
 
-	<div class="two-col">
+	<div class="two-col has-2-columns" style="max-width: none;">
 
-		<div class="col">
+		<div class="col column">
 
 			<h2><strong><?php esc_html_e( 'Do you like this theme?', 'reykjavik' ); ?></strong></h2>
 
@@ -37,7 +37,7 @@
 				(<a href="https://www.webmandesign.eu" target="_blank"><?php esc_html_e( 'Go to WebMan Design website &raquo;', 'reykjavik' ); ?></a>)
 				<?php esc_html_e( 'Or perhaps you are considering a small donation?', 'reykjavik' ); ?>
 				&rarr;
-				<a href="https://www.webmandesign.eu/contact/?utm_source=reykjavik" target="_blank"><em><?php esc_html_e( '"Hey Oliver, have a gallon of coffee on me :)"', 'reykjavik' ); ?></em></a>
+				<a href="https://webmandesign.eu/contact/?utm_source=reykjavik" target="_blank"><em><?php esc_html_e( '"Hey Oliver, have a gallon of coffee on me :)"', 'reykjavik' ); ?></em></a>
 			</p>
 
 			<p>
@@ -48,7 +48,7 @@
 			</p>
 
 			<p>
-				<a href="https://www.webmandesign.eu/contact/?utm_source=reykjavik" target="_blank" class="welcome-upgrade-button"><?php esc_html_e( 'Support theme development', 'reykjavik' ); ?></a>
+				<a href="https://webmandesign.eu/contact/?utm_source=reykjavik" target="_blank" class="welcome-upgrade-button"><?php esc_html_e( 'Support theme development', 'reykjavik' ); ?></a>
 			</p>
 
 			<p class="welcome-upgrade-thanks">
@@ -57,7 +57,7 @@
 
 		</div>
 
-		<div class="col">
+		<div class="col column">
 
 			<h2><strong><?php esc_html_e( 'Feel like stepping up?', 'reykjavik' ); ?></strong></h2>